requirements.txt
langchain
langchain-community
langchain-groq
langchain-huggingface
faiss-cpu
sentence-transformers
python-dotenv
langchainhub
